feat: projen settings
