+++ .eslintrc.json
@@ -145,7 +145,8 @@
       {
         "devDependencies": [
           "**/test/**",
-          "**/build-tools/**"
+          "**/build-tools/**",
+          "function/src/authorizer.lambda.ts"
         ],
         "optionalDependencies": false,
         "peerDependencies": true
