+++ .npmignore
@@ -1,0 +1,27 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+/function/
+!/assets/function/
+/.projen/
+/test-reports/
+junit.xml
+/coverage/
+permissions-backup.acl
+/dist/changelog.md
+/dist/version.txt
+/.mergify.yml
+/test/
+/tsconfig.dev.json
+/src/
+!/lib/
+!/lib/**/*.js
+!/lib/**/*.d.ts
+dist
+/tsconfig.json
+/.github/
+/.vscode/
+/.idea/
+/.projenrc.js
+tsconfig.tsbuildinfo
+/.eslintrc.json
+!.jsii
+!/assets/
