+++ .projen/files.json
@@ -13,6 +13,7 @@
     ".projen/deps.json",
     ".projen/files.json",
     ".projen/tasks.json",
+    "function/src/authorizer-function.ts",
     "LICENSE",
     "tsconfig.dev.json"
   ],
