+++ function/.gitattributes
@@ -1,0 +1,5 @@
+# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+
+/.gitattributes linguist-generated
+/.projen/** linguist-generated
+/yarn.lock linguist-generated
