+++ package.json
@@ -1,0 +1,156 @@
+{
+  "name": "@cloudy-with-a-chance-of-meatballs/cdk-lambda-token-authorizer-jwt",
+  "description": "Add a lambda function to your aws-rest-api-gateway which can be used as a token authorizer",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/cloudy-with-a-chance-of-meatballs/cdk-lambda-token-authorizer-jwt.git"
+  },
+  "scripts": {
+    "build": "npx projen build",
+    "bump": "npx projen bump",
+    "bundle": "npx projen bundle",
+    "bundle:function/src/authorizer.lambda": "npx projen bundle:function/src/authorizer.lambda",
+    "bundle:function/src/authorizer.lambda:watch": "npx projen bundle:function/src/authorizer.lambda:watch",
+    "clobber": "npx projen clobber",
+    "compat": "npx projen compat",
+    "compile": "npx projen compile",
+    "default": "npx projen default",
+    "docgen": "npx projen docgen",
+    "eject": "npx projen eject",
+    "eslint": "npx projen eslint",
+    "package": "npx projen package",
+    "package-all": "npx projen package-all",
+    "package:dotnet": "npx projen package:dotnet",
+    "package:js": "npx projen package:js",
+    "package:python": "npx projen package:python",
+    "post-compile": "npx projen post-compile",
+    "pre-compile": "npx projen pre-compile",
+    "release": "npx projen release",
+    "test": "npx projen test",
+    "test:watch": "npx projen test:watch",
+    "unbump": "npx projen unbump",
+    "watch": "npx projen watch",
+    "projen": "npx projen",
+    "cdk": "cdk"
+  },
+  "author": {
+    "name": "cfuerst",
+    "email": "c.fuerst@gmail.com",
+    "organization": false
+  },
+  "devDependencies": {
+    "@types/jest": "^27",
+    "@types/node": "^16",
+    "@typescript-eslint/eslint-plugin": "^5",
+    "@typescript-eslint/parser": "^5",
+    "esbuild": "^0.15.15",
+    "eslint": "^8",
+    "eslint-import-resolver-node": "^0.3.6",
+    "eslint-import-resolver-typescript": "^3.5.2",
+    "eslint-plugin-import": "^2.26.0",
+    "jest": "^27",
+    "jest-junit": "^13",
+    "jsii": "^1.71.0",
+    "jsii-diff": "^1.71.0",
+    "jsii-docgen": "^7.0.159",
+    "jsii-pacmak": "^1.71.0",
+    "json-schema": "^0.4.0",
+    "projen": "^0.65.34",
+    "standard-version": "^9",
+    "ts-jest": "^27",
+    "typescript": "^4.9.3"
+  },
+  "peerDependencies": {
+    "aws-cdk-lib": "^2.5.0",
+    "constructs": "^10.0.5"
+  },
+  "dependencies": {
+    "aws-cdk-lib": "^2.51.1",
+    "constructs": "^10.1.169"
+  },
+  "keywords": [
+    "api",
+    "apigateway",
+    "authorizer",
+    "aws",
+    "cdk",
+    "jwks",
+    "jwt",
+    "lambda",
+    "rest",
+    "token",
+    "tokenauthorizer"
+  ],
+  "engines": {
+    "node": ">= 16.18.1"
+  },
+  "main": "lib/index.js",
+  "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
+  "version": "0.0.0",
+  "jest": {
+    "testMatch": [
+      "<rootDir>/src/**/__tests__/**/*.ts?(x)",
+      "<rootDir>/(test|src)/**/*(*.)@(spec|test).ts?(x)"
+    ],
+    "clearMocks": true,
+    "collectCoverage": true,
+    "coverageReporters": [
+      "json",
+      "lcov",
+      "clover",
+      "cobertura",
+      "text"
+    ],
+    "coverageDirectory": "coverage",
+    "coveragePathIgnorePatterns": [
+      "/node_modules/"
+    ],
+    "testPathIgnorePatterns": [
+      "/node_modules/"
+    ],
+    "watchPathIgnorePatterns": [
+      "/node_modules/"
+    ],
+    "reporters": [
+      "default",
+      [
+        "jest-junit",
+        {
+          "outputDirectory": "test-reports"
+        }
+      ]
+    ],
+    "preset": "ts-jest",
+    "globals": {
+      "ts-jest": {
+        "tsconfig": "tsconfig.dev.json"
+      }
+    }
+  },
+  "types": "lib/index.d.ts",
+  "stability": "experimental",
+  "jsii": {
+    "outdir": "dist",
+    "targets": {
+      "python": {
+        "distName": "cloudy-with-a-chance-of-meatballs.cdk-lambda-token-authorizer-jwt",
+        "module": "cloudy_with_a_chance_of_meatballs_cdk_lambda_token_authorizer_jwt"
+      },
+      "dotnet": {
+        "namespace": "CloudyWithAchanceOfMeatballs.CdkLambdaTokenAuthorizerJwt",
+        "packageId": "CloudyWithAchanceOfMeatballs.CdkLambdaTokenAuthorizerJwt"
+      }
+    },
+    "tsc": {
+      "outDir": "lib",
+      "rootDir": "src"
+    }
+  },
+  "resolutions": {
+    "@types/prettier": "2.6.0"
+  },
+  "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
+}
